Split every upgradeability approach in a project itself
