+++ README.md
@@ -1,8 +1,18 @@
-**zeppelin_os labs** is a space for the community to interact, exchange ideas, and get involved in the development of the platform as soon as possible.
+# zeppelin_os labs
 
+## Description
+
+**zeppelin_os labs** is a space for the community to interact, exchange ideas, and get involved in the development of the platform as soon as possible.
 
 Over time, we'll be introducing various components of the platform to our Github repo. We invite you to collaborate on how zeppelin_os will be designed and function. To participate, you can submit thoughts, ideas, and proposals in the form of pull requests and Github issues.
 
 [Read the announcement on the zeppelin_os blog](https://blog.zeppelinos.org/announcing-zeppelin_os-labs/).
 
 Chat with the team and community on our [Slack channel](https://slack.openzeppelin.org/).
+
+## Structure
+
+This repository holds many strategies, approaches and study cases that we are investigating and developing in favor of zeppelin_os.
+
+Each experiment should be a project itself so we can make sure their contracts and configurations don't conflict between each other.
+
