+++ upgradeability_using_generic_eternal_storage/truffle.js
@@ -1,0 +1,5 @@
+module.exports = {
+  networks: {
+    // add network configurations here
+  }
+}
