+++ upgradeability_using_inherited_storage/contracts/IRegistry.sol
